msm: dsi-staging: dsi_display: change some var to `u32`
*fixes compilation error

../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
  237 |                 bl_temp = (u32)(bl_dcycle_level + 0.5);
      |                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/gpu/drm/msm/dsi-staging/dsi_display.c:237:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
make[5]: *** [../scripts/Makefile.build:362: drivers/gpu/drm/msm/dsi-staging/dsi_display.o] Error 1
make[4]: *** [../scripts/Makefile.build:676: drivers/gpu/drm/msm] Error 2
make[3]: *** [../scripts/Makefile.build:676: drivers/gpu/drm] Error 2
make[2]: *** [../scripts/Makefile.build:676: drivers/gpu] Error 2
make[1]: *** [/mnt/kernel_dev/violet_kernel/oem/Makefile:1120: drivers] Error 2
make: *** [Makefile:146: sub-make] Error 2

+++ drivers/gpu/drm/msm/dsi-staging/dsi_display.c
@@ -198,8 +198,8 @@ int dsi_display_set_backlight(struct drm_connector *connector,
 	u32 bl_scale, bl_scale_ad;
 	u64 bl_temp;
 	int rc = 0;
-	float bl_dcycle_level;
-	float bl_dcycle = 0.7;
+	unsigned int bl_dcycle_level;
+	unsigned int bl_dcycle = 0.7;
 
 	if (dsi_display == NULL || dsi_display->panel == NULL)
 		return -EINVAL;
